Changed note to notes in documentation for each module
Removed a printline statement from hyper fine parser to streamline output

made a new module OrbitalEnergyParser.py which should eventually parse out the orbital levels for spinup and spin down electrons. Currently not working as intended. But almost.

added orbital energies to the list of things that you can select in main.

Clarified the documentation in SinglePointEnergyParser.py to make it clear what is going on.